+++ R/create_beta_correction_plot_functions.R
@@ -341,10 +341,7 @@ beta_corrected_scatter <- function(varname, q = 0, color = NULL, marginal = FALS
 #'
 #' explain_objects <- explain_iblm(iblm_model, df_list$test)
 #'
-#' # plot can be for a categorical variable (produces list of plots, one for each level)
-#' explain_objects$beta_corrected_density(varname = "Area")
-#'
-#' # plot can be for a single categorical level
+#' # plot can be for a single categorical level OR a categorical variable
 #' explain_objects$beta_corrected_density(varname = "AreaB")
 #'
 #' # output can be numerical variable