+++ R/explain_iblm.R
@@ -204,15 +204,7 @@ data_to_onehot <- function(data, iblm_model, remove_target = TRUE) {
 #'   family = "poisson"
 #' )
 #'
-#' shap <- stats::predict(
-#'   iblm_model$booster_model,
-#'   newdata = xgboost::xgb.DMatrix(
-#'     data.matrix(
-#'       dplyr::select(df_list$test, -dplyr::all_of(iblm_model$response_var))
-#'     )
-#'   ),
-#'   predcontrib = TRUE
-#' ) |> data.frame()
+#' shap <- extract_booster_shap(iblm_model$booster_model, df_list$test)
 #'
 #' wide_input_frame <- data_to_onehot(df_list$test, iblm_model)
 #'
@@ -241,8 +233,7 @@ shap_to_onehot <- function(shap,
     cat_frame <- lapply(names(levels_all_cat), function(x) {
       lvl <- levels_all_cat[[x]]
       mask <- wide_input_frame |>
-        dplyr::select(dplyr::all_of(paste0(x, lvl))) |>
-        data.matrix()
+        dplyr::select(dplyr::all_of(paste0(x, lvl)))
       matrix(rep(shap[, x], length(lvl)), byrow = FALSE, ncol = length(lvl)) * mask
     }) |>
       dplyr::bind_cols()
@@ -288,15 +279,7 @@ shap_to_onehot <- function(shap,
 #'   family = "poisson"
 #' )
 #'
-#' shap <- stats::predict(
-#'   iblm_model$booster_model,
-#'   newdata = xgboost::xgb.DMatrix(
-#'     data.matrix(
-#'       dplyr::select(df_list$test, -dplyr::all_of(iblm_model$response_var))
-#'     )
-#'   ),
-#'   predcontrib = TRUE
-#' ) |> data.frame()
+#' shap <- extract_booster_shap(iblm_model$booster_model, df_list$test)
 #'
 #' wide_input_frame <- data_to_onehot(df_list$test, iblm_model)
 #'
@@ -349,7 +332,7 @@ beta_corrections_derive <- function(shap_wide,
         )
       )
   } else {
-    shap_for_cat_ref <- 0
+    shap_for_cat_ref <- rep(0, nrow(beta_corrections))
   }
 
   beta_corrections$bias <- beta_corrections$bias + shap_for_zeros + shap_for_cat_ref